# docs/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/schedules": {
      "get": {
        "operationId": "ScheduleController_list",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 25,
              "type": "number"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Serbest metin arama",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "institutionId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "groupId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "teacherId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "courseId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "dayOfWeek",
            "required": false,
            "in": "query",
            "description": "0 = Pazartesi ... 6 = Pazar",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "isActive",
            "required": false,
            "in": "query",
            "schema": {
              "default": true,
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": ["schedules"],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "post": {
        "operationId": "ScheduleController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateScheduleDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": ["schedules"],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/schedules/{id}": {
      "get": {
        "operationId": "ScheduleController_get",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": ["schedules"],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "patch": {
        "operationId": "ScheduleController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateScheduleDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": ["schedules"],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "delete": {
        "operationId": "ScheduleController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "tags": ["schedules"],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "Öğrenci Takip Sistemi API",
    "description": "",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateScheduleDto": {
        "type": "object",
        "properties": {
          "groupId": {
            "type": "string"
          },
          "courseId": {
            "type": "string"
          },
          "teacherId": {
            "type": "string",
            "description": "Grup burs programliysa hoca o yurt+programa gorevlendirilmis olmali"
          },
          "dayOfWeek": {
            "type": "number",
            "description": "0 = Pazartesi ... 6 = Pazar"
          },
          "startTime": {
            "type": "string",
            "example": "18:00"
          },
          "endTime": {
            "type": "string",
            "example": "19:30"
          },
          "classroom": {
            "type": "string",
            "example": "A-101"
          },
          "weeklyFrequency": {
            "type": "number",
            "default": 1,
            "description": "Bilgi amacli; haftada iki ders = iki ayri kayit"
          }
        },
        "required": ["groupId", "courseId", "teacherId", "dayOfWeek", "startTime", "endTime"]
      },
      "UpdateScheduleDto": {
        "type": "object",
        "properties": {
          "courseId": {
            "type": "string"
          },
          "teacherId": {
            "type": "string"
          },
          "dayOfWeek": {
            "type": "number"
          },
          "startTime": {
            "type": "string"
          },
          "endTime": {
            "type": "string"
          },
          "classroom": {
            "type": "object",
            "nullable": true
          },
          "weeklyFrequency": {
            "type": "number"
          },
          "isActive": {
            "type": "boolean"
          }
        }
      },
...
    }
  }
}